Move node to graph directory
